displaying user profile, connecting to github and saving data in cookies

+++ server_back/src/routes/memories.ts
@@ -3,9 +3,21 @@ import { prisma } from "../lib/prisma";
 import { z } from 'zod'
  
 export async function memoriesRoutes(app: FastifyInstance) {
-    
-    app.get('/memories', async() => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+
+    app.get('/memories', async(request) => {
+
+        await request.jwtVerify() //verificaçã user
+
+        //retornar apenas memórias do user logado
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -20,7 +32,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }); 
 
       
-      app.get('/memories/:id', async(request) => {
+      app.get('/memories/:id', async(request, reply) => {
         //const { id } = request.params
 
         const paramsSchema= z.object({
@@ -34,6 +46,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             }
         })
+
+        if (!memory.isPublic && memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
 
         return memory
 
@@ -53,7 +69,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '8330d993-e3ec-4f95-8d81-e37e5d9b7c6c',
+                userId: request.user.sub,
             },
         })
         return memory
@@ -61,7 +77,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       
 
 
-      app.put('/memories/:id', async(request) => {
+      app.put('/memories/:id', async(request, reply) => {
 
         const paramsSchema= z.object({
             id: z.string().uuid(),
@@ -78,7 +94,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
       
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
+
+
+         memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -94,12 +121,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
 
 
-      app.delete('/memories/:id', async(request) => {
+      app.delete('/memories/:id', async(request, reply) => {
         const paramsSchema= z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
